# todo_api/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "todo_api, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/Task/GetTaskByGroup": {
      "get": {
        "tags": [
          "Task"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "userid",
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "query",
            "name": "groupId",
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Task/SearchTask": {
      "get": {
        "tags": [
          "Task"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "userid",
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "query",
            "name": "title",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Task/GetTaskIscomplete": {
      "get": {
        "tags": [
          "Task"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "userid",
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Task/GetTaskIsMyDay": {
      "get": {
        "tags": [
          "Task"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "userid",
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Task/GetTaskIsImportant": {
      "get": {
        "tags": [
          "Task"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "userid",
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Task/GetTaskByUser": {
      "get": {
        "tags": [
          "Task"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "userid",
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Task": {
      "get": {
        "tags": [
          "Task"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Task/add": {
      "post": {
        "tags": [
          "Task"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/DTOTask"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Task/UpdateTitle": {
      "put": {
        "tags": [
          "Task"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "query",
            "name": "newTitle",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Task/UpdateDescription": {
      "put": {
        "tags": [
          "Task"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "query",
            "name": "newDes",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Task/UpdateStatusImportant": {
      "put": {
        "tags": [
          "Task"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "query",
            "name": "StatusImportant",
            "type": "boolean"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Task/UpdateStatusMyDay": {
      "put": {
        "tags": [
          "Task"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "query",
            "name": "StatusMyDay",
            "type": "boolean"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Task/UpdateStatusComplete": {
      "put": {
        "tags": [
          "Task"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "query",
            "name": "StatusComplete",
            "type": "boolean"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Task/Delete": {
      "delete": {
        "tags": [
          "Task"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/getAllUser": {
      "get": {
        "tags": [
          "User"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/GetUserByEmail": {
      "get": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "email",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Login": {
      "post": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "email",
            "type": "string"
          },
          {
            "in": "query",
            "name": "password",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Register": {
      "post": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "Username",
            "type": "string"
          },
          {
            "in": "query",
            "name": "Password",
            "type": "string"
          },
          {
            "in": "query",
            "name": "Email",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/UpdateUser": {
      "put": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "Username",
            "type": "string"
          },
          {
            "in": "query",
            "name": "Password",
            "type": "string"
          },
          {
            "in": "query",
            "name": "Email",
            "type": "string"
          },
          {
            "in": "query",
            "name": "id",
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/WeatherForecast": {
      "get": {
        "tags": [
          "WeatherForecast"
        ],
        "operationId": "GetWeatherForecast",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/WeatherForecast"
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
    "DTOTask": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "userId": {
          "format": "uuid",
          "type": "string"
        },
        "isCompleted": {
          "type": "boolean"
        },
        "isImportant": {
          "type": "boolean"
        },
        "isMyDay": {
          "type": "boolean"
        },
        "groupId": {
          "format": "uuid",
          "type": "string"
        },
        "dueDate": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
    "DateOnly": {
      "type": "object",
      "properties": {
        "year": {
          "format": "int32",
          "type": "integer"
        },
        "month": {
          "format": "int32",
          "type": "integer"
        },
        "day": {
          "format": "int32",
          "type": "integer"
        },
        "dayOfWeek": {
          "$ref": "#/definitions/DayOfWeek"
        },
        "dayOfYear": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "dayNumber": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        }
      }
    },
    "DayOfWeek": {
      "format": "int32",
      "enum": [
        0,
        1,
        2,
        3,
        4,
        5,
        6
      ],
      "type": "integer"
    },
    "WeatherForecast": {
      "type": "object",
      "properties": {
        "date": {
          "$ref": "#/definitions/DateOnly"
        },
        "temperatureC": {
          "format": "int32",
          "type": "integer"
        },
        "temperatureF": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "summary": {
          "type": "string"
        }
      }
    }
  }
}